Module Tezos_protocol_environment

module Context : sig ... end

module Register (C : sig ... end) : sig ... end

module Environment_profiler : sig ... end

val err_implementation_mismatch : expected:string -> got:string -> 'a

type validation_result = {

context : Context.t;

fitness : Tezos_base.TzPervasives.Fitness.t;

message : string option;

max_operations_ttl : int;

last_finalized_block_level : Stdlib.Int32.t;

last_preserved_block_level : Stdlib.Int32.t;

}

type quota = {

max_size : int;

max_op : int option;

}

type rpc_context = {

block_hash : Tezos_crypto.Hashed.Block_hash.t;

block_header : Tezos_base.TzPervasives.Block_header.shell_header;

context : Context.t;

}

module type PROTOCOL = sig ... end

module V0 : sig ... end

module V1 : sig ... end

module V2 : sig ... end

module V3 : sig ... end

module V4 : sig ... end

module V5 : sig ... end

module V6 : sig ... end

module V7 : sig ... end

module V8 : sig ... end

module V9 : sig ... end

module V10 : sig ... end

module V11 : sig ... end

module V12 : sig ... end

module V13 : sig ... end

module V14 : sig ... end

module Memory_context : sig ... end

Memory_context is a recursive map backed by an in-memory Irmin tree, i.e. Tezos_context_memory.Context (in src/lib_context), used by the mockup mode and by proof verifiers (i.e. clients of Tezos_context_helpers.Context.Make_tree.Proof). It abstracts away Tezos_context_memory.Context.t.

module Brassaia_memory_context : sig ... end

module Proxy_context : sig ... end

This module is the location where the proxy tweaks the behavior of a vanilla client. A regular mockup client uses a Memory_context in place of this implementation. Compared to Memory_context, this instance features a Proxy_Delegate.T which under the hood relies on Proxy_getter.

module Proxy_delegate : sig ... end

module Internal_for_tests : sig ... end