Module Environment_profiler.Environment_profiler

type nonrec metadata = Tezos_base.Profiler.metadata

type nonrec id = Tezos_base.Profiler.id

type nonrec ids = Tezos_base.Profiler.ids

type nonrec verbosity = Tezos_base.Profiler.verbosity =

| Notice

| Info

| Debug

Plugging

val plug : Tezos_base.Profiler.instance -> unit

plug instance plugs the profiler stored in the module to instance

val unplug : Tezos_base.Profiler.instance -> unit

unplug instance unplugs the profiler stored in the module from instance

val close_and_unplug : Tezos_base.Profiler.instance -> unit

close_and_unplug instance closes the profiler stored in the module and unplugs it from instance

val close_and_unplug_all : unit -> unit

close_and_unplug_all () closes the profiler stored in the module and unplugs it from all instances it was plugged to

val plugged : unit -> Tezos_base.Profiler.instance list

plugged () returns all the instances plugged to the profiler stored in the module

Sequences

val record : verbosity -> id -> unit

Open a sequence in the current sequence. If currently aggregating (not all aggregation scopes are closed), this has the same semantics as aggregate instead.

val aggregate : verbosity -> id -> unit

Open an aggregation node in the current sequence.

val stop : unit -> unit

Close the most recently opened sequence or aggregation scope.

Profiling

val stamp : verbosity -> id -> unit

Record a timestamp in the most recently opened sequence.

val mark : verbosity -> ids -> unit

Count this event's occurences in the most recent sequence.

val span : verbosity -> Tezos_base.Profiler.span -> ids -> unit

Sum the time spent in this event in the most recent sequence.

val inc : Tezos_base.Profiler.report -> unit

Include a report in the current sequence.

val record_f : verbosity -> id -> (unit -> 'a) -> 'a

record_f verbosity label f will call:

record verbosity name;
f ();
stop ();

val record_s : verbosity -> id -> (unit -> 'a Lwt.t) -> 'a Lwt.t

Same as record_f but for Lwt function

val aggregate_f : verbosity -> id -> (unit -> 'a) -> 'a

aggregate_f verbosity label f will call:

aggregate verbosity name;
f ();
stop ();

val aggregate_s : verbosity -> id -> (unit -> 'a Lwt.t) -> 'a Lwt.t

Same as aggregate_f but for Lwt functions

val span_f : verbosity -> ids -> (unit -> 'a) -> 'a

span_f verbosity label_list f will compute span but specifically around f

val span_s : verbosity -> ids -> (unit -> 'a Lwt.t) -> 'a Lwt.t

Same as span_f but for Lwt functions