Module Internal_for_tests.Environment_protocol_T_test

module Mock_all_unit : sig ... end

An implementation of Tezos_protocol_environment.PROTOCOL used in tests. It sets all types to unit and implement all functions as assert false. Users of this module should include it and override the functions they need.