Module type Proxy_delegate.T

The module that needs to be implemented for providing the delegation feature to Proxy_context. Implementors willing to add a new backend to the client's --mode proxy should likely add a new implementation of this module type.

val proxy_dir_mem : string list -> bool Tezos_base.TzPervasives.tzresult Lwt.t

proxy_dir_mem key returns whether proxy_get would return a non-leaf tree.

val proxy_get : string list -> Tezos_context_memory.Context.tree option Tezos_base.TzPervasives.tzresult Lwt.t

proxy_get key returns the tree associated to key.

val proxy_mem : string list -> bool Tezos_base.TzPervasives.tzresult Lwt.t

proxy_mem key returns whether proxy_get key would return a leaf tree.