Module T.Context

View over the context store, restricted to types, access and functional manipulation of an existing context.

type t

type key = string list

Keys in (kex x value) database implementations

type value = bytes

Values in (kex x value) database implementations

val mem : t -> key -> bool Lwt.t

val dir_mem : t -> key -> bool Lwt.t

val get : t -> key -> value option Lwt.t

val set : t -> key -> value -> t Lwt.t

val copy : t -> from:key -> to_:key -> t option Lwt.t

copy returns None if the from key is not bound

val remove_rec : t -> key -> t Lwt.t

type key_or_dir = [

| `Key of key

| `Dir of key

]

val fold : t -> key -> init:'a -> f:(key_or_dir -> 'a -> 'a Lwt.t) -> 'a Lwt.t

val keys : t -> key -> key list Lwt.t

val fold_keys : t -> key -> init:'a -> f:(key -> 'a -> 'a Lwt.t) -> 'a Lwt.t

val register_resolver : 'a Base58.encoding -> (t -> string -> 'a list Lwt.t) -> unit

val complete : t -> string -> string list Lwt.t