Module type Bounded.BOUNDS

Bounds.

Formally each B : BOUND represents the interval of all values between B.min_value and B.max_value. This is a closed interval, i.e. the endpoints are included.

Intervals can be empty, for example struct let min_value = 1; let max_value 0 end is empty.

type ocaml_type

ocaml_type is the type used for the internal representation of values within the bounded interval. This is the type that values in the interval are converted to and from. E.g., for an interval of 32-bit integers ocaml_type = int32.

val min_value : ocaml_type

min_value represents the minimal value (included) reprensatable.

val max_value : ocaml_type

max_value represents the maximal value (included) reprensatable.