Module T.Array

This module is a very restricted subset of OCaml's Stdlib Array module. There is just enough exposed that you can pass arrays around to some functions and such, but not enough that you can easily shoot yourself in the foot. (See details below.)

If you need Arrays with more features, you should check the FallbackArray module.

type 'a t = 'a array

The type of native OCaml arrays. You can construct them with the literal syntax ([|"like"; "so"|]) or obtain them by deserialising data.

val concat : 'a t list -> 'a t

val length : 'a t -> int

val to_list : 'a t -> 'a list