Module T.Fitness

The fitness of a block is defined as a list of bytes, compared in a lexicographical order (longer list are greater).

include S.T with type t = bytes list

type t = bytes list

include Compare.S with type t := t

val (=) : t -> t -> bool

x = y iff compare x y = 0

val (<>) : t -> t -> bool

x <> y iff compare x y <> 0

val (<) : t -> t -> bool

x < y iff compare x y < 0

val (<=) : t -> t -> bool

x <= y iff compare x y <= 0

val (>=) : t -> t -> bool

x >= y iff compare x y >= 0

val (>) : t -> t -> bool

x > y iff compare x y > 0

val compare : t -> t -> int

compare an alias for the functor parameter's compare function

val equal : t -> t -> bool

equal x y iff compare x y = 0

val max : t -> t -> t

max x y is x if x >= y otherwise it is y

val min : t -> t -> t

min x y is x if x <= y otherwise it is y

val pp : Format.formatter -> t -> unit

val encoding : t Data_encoding.t

val to_bytes : t -> bytes

val of_bytes : bytes -> t option