Module Tezos_protocol_environment_structs.V4

module Data_encoding : sig ... end

module Replicated_signatures : sig ... end

We have to duplicate the content of src/lib_protocol_environment/sigs/v3/set.mli (and a few other signatures from the same directory) so we can refer to them in src/lib_protocol_environment/environment_V3.ml. Indeed, it is impossible to refer to components of Tezos_protocol_environment_sigs.V3.T (such as the Set component) because T is a module type (not a module).

module Lwtreslib_list_combine : sig ... end

module Error_monad_infix_globals : sig ... end

module RPC_directory : sig ... end

module Seq : sig ... end