Module type Int32.BOUNDS

Bounds.

Formally each B : BOUND represents the interval of all integers between B.min_int and B.max_int. This is a closed interval, e.g. the endpoints are included.

Intervals can be empty, for example struct let min_int = 1; let max_int 0 end is empty.

val min_int : int32

val max_int : int32