Module Encoding.Bounded

val string' : ?length_kind:[ `N | `Uint30 | `Uint16 | `Uint8 ] -> Data_encoding__V1.Encoding.string_json_repr -> int -> string Data_encoding.Encoding.t

Encoding of a string whose length does not exceed the specified length.

If length_kind is set, then it is used to encode the length of the string in a header. If length_kind is omitted then the length field uses the smallest fixed-width integer that can accommodate the maximum size - e.g., `Uint8 for very short strings, `Uint16 for longer strings, etc.

Attempting to construct a string with a length that is too long causes an Invalid_argument exception.

raises Invalid_argument
if length_kind is set but it cannot accommodate the specified bound. E.g., Bounded.string' ~length_kind:`Uint8 Hex 1000 raises.

raises Invalid_argument
if specified bound is negative or null.

raises Invalid_argument
if length_kind is unset and the specified bound is larger than 2^30.

val bytes' : ?length_kind:[ `N | `Uint30 | `Uint16 | `Uint8 ] -> Data_encoding__V1.Encoding.string_json_repr -> int -> Stdlib.Bytes.t Data_encoding.Encoding.t

See string' above.

val bigstring : ?length_kind:[ `N | `Uint30 | `Uint16 | `Uint8 ] -> ?string_json_repr:Data_encoding__V1.Encoding.string_json_repr -> int -> (char, Stdlib.Bigarray.int8_unsigned_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t Data_encoding.Encoding.t

Same as string but for bigstring.

parameter [?string_json_repr]
defaults to Hex.

raises Invalid_argument
if length_kind is set but it cannot accommodate the specified bound. E.g., Bounded.string' ~length_kind:`Uint8 Hex 1000 raises.

raises Invalid_argument
if specified bound is negative or null.

val string : ?length_kind:[ `N | `Uint16 | `Uint30 | `Uint8 ] -> Data_encoding__V1.Encoding.string_json_repr -> int -> string Data_encoding.Encoding.t

val bytes : ?length_kind:[ `N | `Uint16 | `Uint30 | `Uint8 ] -> Data_encoding__V1.Encoding.string_json_repr -> int -> Stdlib.Bytes.t Data_encoding.Encoding.t