Module Encoding.Variable

Create encodings that produce data of a variable length when binary encoded. See the preamble for an explanation.

val string' : Data_encoding__V1.Encoding.string_json_repr -> string Data_encoding.Encoding.t

val bytes' : Data_encoding__V1.Encoding.string_json_repr -> Stdlib.Bytes.t Data_encoding.Encoding.t

val bigstring : ?string_json_repr:Data_encoding__V1.Encoding.string_json_repr -> unit -> (char, Stdlib.Bigarray.int8_unsigned_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t Data_encoding.Encoding.t

val array : ?max_length:int -> 'a Data_encoding.Encoding.t -> 'a array Data_encoding.Encoding.t

raises Invalid_argument
if the encoding argument is variable length or may lead to zero-width representation in binary.

val list : ?max_length:int -> 'a Data_encoding.Encoding.t -> 'a list Data_encoding.Encoding.t

raises Invalid_argument
if the encoding argument is variable length or may lead to zero-width representation in binary.

val string : Data_encoding__V1.Encoding.string_json_repr -> string Data_encoding.Encoding.t

val bytes : Data_encoding__V1.Encoding.string_json_repr -> Stdlib.Bytes.t Data_encoding.Encoding.t